refactor(spinner): extract spinner detection/collapse to output-cap.ts
- Move SPINNER_CHARS (braille + quarter-circle; ASCII chars removed to
  avoid false positives on table separators, diff/graph markers, and
  line continuations) and collapseSpinnerChunks into output-cap.ts
- ToolOverlayLog.tsx and PersistentShellBlock.tsx import the shared
  utility instead of duplicating the detection set and loop
- Fixes SPINNER_CHARS being declared in the middle of PersistentShellBlock's
  import block (P2 position issue)

Addresses three P2 comments from ReAgent review on PR #1716.
