+++ frontend/app/view/agent/components/PersistentShellBlock.tsx
@@ -13,14 +13,7 @@
 
 import clsx from "clsx";
 import { For, Show, createEffect, createMemo, createSignal, onCleanup, type JSX } from "solid-js";
-import { capChars, createChunkCapper, MAX_TOOL_OUTPUT_LINES } from "./output-cap";
-
-const SPINNER_CHARS = new Set([
-    '⠋','⠙','⠹','⠸','⠼','⠴','⠦','⠧','⠇','⠏',
-    '⣾','⣽','⣻','⢿','⡿','⣟','⣯','⣷',
-    '◐','◓','◑','◒','◴','◷','◶','◵',
-    '-','\\','|','/',
-]);
+import { capChars, collapseSpinnerChunks, createChunkCapper, MAX_TOOL_OUTPUT_LINES } from "./output-cap";
 import { OutputHiddenMarker } from "./OutputHiddenMarker";
 import { LinkifiedText } from "@/app/element/linkified-text";
 import { RpcApi } from "@/app/store/rpc-api";
@@ -108,31 +101,7 @@ export const PersistentShellBlock = (props: PersistentShellBlockProps): JSX.Elem
     const chunkCap = createChunkCapper(MAX_TOOL_OUTPUT_LINES);
     const cappedView = createMemo(() => {
         const { chunks, hiddenLines } = chunkCap(props.node.log.chunks as ToolLogChunk[]);
-        // Collapse consecutive spinner-char runs: trailing run → live slot,
-        // completed run → frozen last frame. Mirrors ToolOverlayLog ChunkList.
-        const display: ToolLogChunk[] = [];
-        let spinnerSlot: { content: string; kind: string } | null = null;
-        for (let i = 0; i < chunks.length; i++) {
-            const chunk = chunks[i];
-            const trimmed = capChars(chunk.content).trim();
-            if (SPINNER_CHARS.has(trimmed)) {
-                let last = chunk;
-                while (i + 1 < chunks.length && SPINNER_CHARS.has(capChars(chunks[i + 1].content).trim())) {
-                    i++;
-                    last = chunks[i];
-                }
-                const lastFrame = capChars(last.content).trim();
-                if (i === chunks.length - 1) {
-                    spinnerSlot = { content: lastFrame, kind: last.kind };
-                } else {
-                    display.push({ ...last, content: lastFrame });
-                    spinnerSlot = null;
-                }
-            } else {
-                display.push(chunk);
-                spinnerSlot = null;
-            }
-        }
+        const { display, spinnerSlot } = collapseSpinnerChunks(chunks);
         return { display, spinnerSlot, hiddenLines };
     });
     const visibleChunks = () => cappedView().display;
